Rustifying the website! And other changes
Rust used as both templating engine for HTML (via Askama) and, in
future, web framework for the website; in order to use Rust instead of
Go. In that matter, these templates allow to statically generate the
website HTML code (i.e. such as the navigation bar for every website)
ready for deployment, without the need to generate any of those elements
within the user's browser through Javascript, which removes such overhead.

Currently, a plain-text file is used to store the categories of which
each show, series, game or manga are in; this file can be parsed for
further context and, in the future, simaultenously generate HTML code
for each item.

[However, Go has to be used to deploy the website, for now].